165864436-chore: remove ativity functionality
